Make sure Descriptor::densify follows librascal convention
In order to match librascal, we want to sort species pairs,
this means using a BTreeSet for new features in remove_from_samples.
Because of this, we can not pre-determine the position of the
feature blocks, and instead we have to lookup the the actual feature
block when moving data around.