Update hyper-parameters of density expansions to a more extensible schema
There are two main ingredients to a spherical expansion:

- the density being expanded
- the basis on which to make the expansion

Additionally, real space/short range (i.e. SOAP) density expansion requires the
definition of local environments with a cutoff radius and smoothing functions.

This commit refactor the hyper-parameters of spherical expansion and related calculators
to reflect this structure and make is easier to add new variations in the density and
basis ingredients.